requirements.txt
Flask==3.0.3
requests
mysql-connector-python
Werkzeug
flask-socketio
groq 
mistralai
sentence-transformers 
chromadb 
pypdf 
pillow 
python-dotenv
Flask-Session